+++ src/lib/components/SecondBar.svelte
@@ -16,16 +16,24 @@
 	let pct = $derived((seconds / 60) * 100);
 </script>
 
-{#if prefs.secondStyle === 'fullscreen-bar'}
-	<div
-		class="pointer-events-none absolute top-0 left-0 h-full transition-none"
-		style="width: {pct}%; background: {prefs.colorProgress}; opacity: {prefs.progressOpacity /
-			100};"
-	></div>
-{:else if prefs.secondStyle === 'top-bar'}
-	<div
-		class="pointer-events-none absolute top-0 left-0 h-12 transition-none"
-		style="width: {pct}%; background: {prefs.colorProgress}; opacity: {prefs.progressOpacity /
-			100};"
-	></div>
+{#if prefs.secondStyle === 'bar'}
+	{#if prefs.barPosition === 'fullscreen'}
+		<div
+			class="pointer-events-none absolute top-0 left-0 h-full transition-none"
+			style="width: {pct}%; background: {prefs.colorProgress}; opacity: {prefs.progressOpacity /
+				100};"
+		></div>
+	{:else if prefs.barPosition === 'top'}
+		<div
+			class="pointer-events-none absolute top-0 left-0 h-12 transition-none"
+			style="width: {pct}%; background: {prefs.colorProgress}; opacity: {prefs.progressOpacity /
+				100};"
+		></div>
+	{:else if prefs.barPosition === 'bottom'}
+		<div
+			class="pointer-events-none absolute bottom-0 left-0 h-12 transition-none"
+			style="width: {pct}%; background: {prefs.colorProgress}; opacity: {prefs.progressOpacity /
+				100};"
+		></div>
+	{/if}
 {/if}